avoid newlines in og:description
In some recent release of Pandoc, intra-para newlines in markdown
parse to SoftBreak in the Pandoc AST.  When rendered as HTML each
SoftBreak turns back into a newline character.  This has no effect
on HTML rendering, but it DOES have an impact in the og:description.
It can result in weird appearance due to unexpected line breaks.

To avoid this, update `removeFormatting` to replace SoftBreak with
Space.