site: update to hakyll-4.16 and fix double-write
Hakyll 4.16.1.0 changed the runtime, which now errors if it writes
the same output file twice.  We have two versions of each post
("recent" for metadata, snapshots and the recent-posts list; and the
no-version that applies the HTML templates), and both had a route,
which triggered the warning.  But removing the route from either
version breaks hyperlinks where the `$url$` template field occurs.

The workaround has two parts.  For the recent posts list, define
`urlFieldNoVersion` that retrieves the route of the no-version
variant of the current identifier.  We add this field with the
`"url0"` and update the `recent.html` template.

For the atom feed, the included templates use the `$url$` field.  We
work around this be re-snapshotting the post content in the
no-version Rules, and loading that no-version snapshot (which has
the route information available).

Finally, I also rearraged some of the Rules and avoided one
unnecessary content snapshot.